% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial.R
\name{distance}
\alias{distance}
\title{find the distance between two points}
\usage{
distance(a, b, x, y, type = "abs")
}
\arguments{
\item{a}{x of first point}

\item{b}{y of first point}

\item{x}{x of first point}

\item{y}{y of first point}
}
\value{
scalar
}
\description{
find the distance between two points
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

