% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial.R
\name{faceplot}
\alias{faceplot}
\title{plot the dots the image}
\usage{
faceplot(data, path, pch = 19, col = "black")
}
\arguments{
\item{data}{dataset contine "x" as x, "y" as y}

\item{path}{path of image}
}
\value{
plot
}
\description{
plot the dots the image
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

