% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial.R
\name{faceplusplus}
\alias{faceplusplus}
\title{Post image to face++ and return face landmark}
\usage{
faceplusplus(img.path, apikey = "i6aAp-TZNec1-HjPlzJS1oAaEcdRUHxi",
  seqkey = "dwobv2uNCA0MvUNTNwIiEmqQEn8jGGLR")
}
\arguments{
\item{img.path}{the path of image}

\item{apikey}{api key}

\item{seqkey}{secrit key}
}
\value{
data.frame
}
\description{
Post image to face++ and return face landmark
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

